pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["MinkyoDev <minkyo.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi-slim = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
fastapi-camelcase = "^2.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
